Added skewed sine wave for Ankle
The left foot now rotates, with the ankle rotating intuitively. More sync is needed between the YZ movement of the foot controller and the Ankle rotation. The next step would be to generalize this for multiple controllers and then focus on refinement. This work is currently inspired by the EMOTE model (Chi 2000).